Move AnchorGenerator to a seperate file.
